+++ spec/on_strum/healthcheck/version_spec.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-RSpec.describe OnStrum::Healthcheck::VERSION do
-  it { is_expected.not_to be_nil }
-end